requirements.txt
Django>=4.0.0
djangorestframework>=3.13.0
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.0
scapy>=2.4.0
requests>=2.27.0
matplotlib>=3.5.0
python-dateutil>=2.8.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.11.0